refactor: move command into dedicated traits namespace
