+++ src/Commands/GetVersion.php
@@ -5,7 +5,7 @@ namespace StellarWP\Pup\Commands;
 use StellarWP\Pup\App;
 use StellarWP\Pup\Exceptions\BaseException;
 use StellarWP\Pup\Command\Command;
-use StellarWP\Pup\Command\DevSuffix;
+use StellarWP\Pup\Command\Traits\DevSuffix;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
